requirements.txt
pydub
argparse
openai-whisper
python-dotenv
